src/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
firebase-admin
scapy
influxdb-client
websockets
python-multipart
asyncio-mqtt
redis
python-dateutil
mitmproxy
python-nmap
pytest
pytest-asyncio
httpx# OpenTelemetry dependencies
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-exporter-prometheus
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-sqlalchemy
structlog